fix: resolve hidden brokenness findings (audit #9) across 3 crates
Addresses 12 findings + 2 refinements from audit #9 code-simplifier review:

- H1: e10_active_vector() falls back to e10_multimodal_as_context for backward compat
- H2: lastHours/lastDays schema changed from number to integer with minimum:1
- M2: PipelineBuilder skips MatryoshkaAnn stage when no matryoshka query provided
- M3: Skip E2/E3/E4 temporal HNSW indexes (never searched, saves CPU+RAM)
- M6: kill_stale results logged at debug level (false=normal on clean startup)
- M7: causal search response returns "direction" field instead of "searchMode"
- M9: cosine_similarity normalized [-1,1]→[0,1] per SRC-3 convention
- L2: RRF_K typed as i32 instead of f32
- L3: expect() instead of unwrap_or_else for infallible stub operations
- L4: PID file removal failures now logged instead of silently ignored
- L5: additionalProperties:false added to 24 MCP tool schemas
- L6: embedder.rs stale comment corrected (5→7 tools)
- Refinement: consolidation threshold 0.85→0.925 (preserves raw cosine 0.85 after SRC-3)

+++ crates/context-graph-storage/src/teleological/rocksdb_store/causal_relationships.rs
@@ -861,7 +861,7 @@ impl RocksDbTeleologicalStore {
                 (e8_results, config.e8_weight, "e8"),
                 (e11_results, config.e11_weight, "e11"),
             ],
-            RRF_K as i32,
+            RRF_K,
         );
 
         // Stage 6: Build results with consensus scoring
